Update patch set 5
Patch Set 5:

This will need to be merged by someone on the current committers list. Which, now that I look at it, is almost entirely Nokia, and there is no overlap with the committers listed in this INFO file.

There are several people still listed on this [0] page that are not in your INFO file, but are on the current committer list.
0: https://wiki.akraino.org/display/AK/Radio+Edge+Cloud+%28REC%29+Team+Membership

Patch-set: 5